Move conjunction and disjunction modules to aql module
