+++ graphannis/src/annis/db/mod.rs
@@ -4,7 +4,6 @@ pub mod corpusstorage;
 pub mod example_generator;
 pub mod exec;
 mod plan;
-pub mod query;
 pub mod relannis;
 pub mod sort_matches;
 pub mod token_helper;
