--- graphannis/src/annis/db/query/mod.rs
@@ -1,7 +1,0 @@
-#[derive(Clone, Default, Debug)]
-pub struct Config {
-    pub use_parallel_joins: bool,
-}
-
-pub mod conjunction;
-pub mod disjunction;
